LLM파트 chatAnthropic으로 수정

+++ app/agents/shift_analyzer_agent.py
@@ -40,7 +40,7 @@ class ShiftSubgraph(TypedDict):
 class shiftResponse(TypedDict):
     shift: str 
     date: List[int] 
-    score: List[int] 
+    score: List[float] 
 
 class shiftAnalyzer(BaseModel):
     processor: str
